Keep largest component

Finds connected components of a graph and returns the largest component

Usage

KeepLargestComponent(object, ...)

# S3 method for tbl_graph
KeepLargestComponent(object, verbose = TRUE, ...)

# S3 method for CellGraph
KeepLargestComponent(object, verbose = TRUE, ...)

# S3 method for CellGraphAssay
KeepLargestComponent(object, verbose = TRUE, ...)

# S3 method for Seurat
KeepLargestComponent(object, assay = NULL, verbose = TRUE, ...)

Arguments

object

An object

...

Parameters passed to other methods

verbose

Print messages

assay

Name of a CellGraphAssay stored on the Seurat object

Examples

library(pixelatorR)
pxl_file <- system.file("extdata/PBMC_10_cells",
                        "Sample01_test.pxl",
                        package = "pixelatorR")

# Read edgelist
edgelist <- ReadMPX_arrow_edgelist(pxl_file, overwrite = TRUE)
#> ℹ Output directory set to '/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T//Rtmp5tX2PT'
#> ℹ Copying parquet files to /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T//Rtmp5tX2PT/y7p6F-2024-02-22-093458
#> ✔ Returning FileSystemDataset

# Load graph from edge list and store in a CellGraph object
cg <- LoadCellGraphs(edgelist, cells = "RCVCMP0000000")[[1]]
#> →   Loading 1 edgelist(s) as bipartite graph(s)
cg
#> A CellGraph object containing a bipartite graph with 3549 nodes and 6909 edges covering 24 markers

# Fetch tbl_graph from CellGraph object
g <- CellGraphData(cg, slot = "cellgraph")
g
#> # A tbl_graph: 3549 nodes and 6909 edges
#> #
#> # An undirected simple graph with 11 components
#> #
#> # Node Data: 3,549 × 2 (active)
#>    name                        node_type
#>    <chr>                       <chr>    
#>  1 ATCAATATCGCTAGCGAAGAACACA-A A        
#>  2 TAGTATCTCGAAGGGATTCGGTGGT-A A        
#>  3 TGATGGTGGCATAATGGTGTATATT-A A        
#>  4 ACAGCGCAATCTCATAGAGGTATAT-A A        
#>  5 GGAAGGTACAGGGCCAACGGAAGTT-A A        
#>  6 TGATCTTGAATTAGGTTGTGTGATA-A A        
#>  7 ATTTCTGTTTAGCACGTGAGCATGA-A A        
#>  8 ATATGGGCGAGTCCAATTGCATATA-A A        
#>  9 ATGTTGGGTATTACAGGACATCGTG-A A        
#> 10 TCCAGGGAGATGTTGGAAGACTTAA-A A        
#> # ℹ 3,539 more rows
#> #
#> # Edge Data: 6,909 × 2
#>    from    to
#>   <int> <int>
#> 1     1  2628
#> 2     1  2710
#> 3     1  2787
#> # ℹ 6,906 more rows

# Fetch largest component from a tbl_graph
g_largest <- KeepLargestComponent(g)
#> ℹ Removed 24 out of 3549 nodes
g_largest
#> # A tbl_graph: 3525 nodes and 6895 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 3,525 × 2 (active)
#>    name                        node_type
#>    <chr>                       <chr>    
#>  1 ATCAATATCGCTAGCGAAGAACACA-A A        
#>  2 TAGTATCTCGAAGGGATTCGGTGGT-A A        
#>  3 TGATGGTGGCATAATGGTGTATATT-A A        
#>  4 ACAGCGCAATCTCATAGAGGTATAT-A A        
#>  5 GGAAGGTACAGGGCCAACGGAAGTT-A A        
#>  6 TGATCTTGAATTAGGTTGTGTGATA-A A        
#>  7 ATTTCTGTTTAGCACGTGAGCATGA-A A        
#>  8 ATATGGGCGAGTCCAATTGCATATA-A A        
#>  9 ATGTTGGGTATTACAGGACATCGTG-A A        
#> 10 TCCAGGGAGATGTTGGAAGACTTAA-A A        
#> # ℹ 3,515 more rows
#> #
#> # Edge Data: 6,895 × 2
#>    from    to
#>   <int> <int>
#> 1     1  2614
#> 2     1  2696
#> 3     1  2773
#> # ℹ 6,892 more rows

# Fetch largest component from a CellGraph
cg_largest <- KeepLargestComponent(cg)
#> ℹ Removed 24 out of 3549 nodes
cg_largest
#> A CellGraph object containing a bipartite graph with 3525 nodes and 6895 edges covering 24 markers