Read edgelists from a PXL file containing MPX data

This function uses arrow to read edgelists from one or several PXL files. The edgelists are stored in parquet files in the outdir directory which can be modified on disk.

Usage

ReadMPX_arrow_edgelist(pxl_file, edge_list_file = NULL, verbose = TRUE, ...)

Arguments

pxl_file

Path to a PXL file

edge_list_file

Path to the output edgelist.parquet file

verbose

Print messages

...

Parameters passed to other methods

Value

Nothing. The edgelist is saved to a parquet file set with edge_list_file

Examples

library(pixelatorR)

# Load example data
pxl_file <- minimal_mpx_pxl_file()
edgelist_arrow <- ReadMPX_arrow_edgelist(pxl_file)
#> ℹ Extracting edgelist.parquet file to /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmplGz5wi/edgelist.parquet
#> ✔ Returning FileSystemDataset
edgelist_arrow
#> FileSystemDataset with 1 Parquet file
#> upia: string
#> upib: string
#> marker: dictionary<values=string, indices=int32>
#> count: int8
#> component: dictionary<values=string, indices=int32>