Export a FileSystemDataset to .parquet files

This function can be used to export a FileSystemDataset to a folder containing .parquet files readable with open_dataset

Usage

export_edgelist_to_parquet(
  object,
  outdir,
  suffix = "",
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

object

An FileSystemDataset

outdir

A path to an existing directory

suffix

Add a suffix to output directory name

overwrite

If the output directory already exists, set this parameter to TRUE to overwrite the directory

verbose

Print messages

Value

Path to the output folder

Examples

library(pixelatorR)
library(dplyr)

# Load example data
pxl_file <- system.file("extdata/PBMC_10_cells",
                        "Sample01_test.pxl",
                        package = "pixelatorR")
edgelist_arrow <- ReadMPX_arrow_edgelist(pxl_file, overwrite = TRUE)
#> ℹ Output directory set to '/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T//RtmpG2lC2V'
#> ℹ Copying parquet files to /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T//RtmpG2lC2V/mYwDD-2024-02-21-113532
#> ✔ Returning FileSystemDataset

# Manipulate edgelist with dplyr verbs
sorted_components <- edgelist_arrow %>%
  group_by(component) %>%
  summarize(component_size = n()) %>%
  arrange(-component_size)

# Inspect results
sorted_components %>% collect()
#> # A tibble: 10 × 2
#>    component     component_size
#>    <chr>                  <int>
#>  1 RCVCMP0000013          14256
#>  2 RCVCMP0000000          14247
#>  3 RCVCMP0000006          12786
#>  4 RCVCMP0000005           7123
#>  5 RCVCMP0000003           4115
#>  6 RCVCMP0000010           3130
#>  7 RCVCMP0000008           3018
#>  8 RCVCMP0000002           2475
#>  9 RCVCMP0000007           2262
#> 10 RCVCMP0000012           1338

# Export results to a parquet file
tmp_dir <- tempdir()
outdir <- file.path(tmp_dir, "sorted_components")
dir.create(outdir, showWarnings = FALSE)
outpath <- export_edgelist_to_parquet(object = sorted_components,
                                      outdir = outdir,
                                      overwrite = TRUE)
#> ℹ Saving edgelist to /private/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpG2lC2V/sorted_components/9o7hY-2024-02-21-113532/
outpath
#> [1] "/private/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpG2lC2V/sorted_components/9o7hY-2024-02-21-113532/"