Global Options in pixelatorR

Information about global options provided in pixelatorR.

Details

Available options:

pixelatorR.arrow_outdir The directory where the edgelists are stored.

pixelatorR.arrowdir_maxsize The maximum allowed size of the edgelists directory.

pixelatorR.auto_cleanup Whether to run edgelists_directories_clean() automatically when the size of the edgelists directory exceeds pixelatorR.arrowdir_maxsize. The cleanup is triggered by functions which create new edgelist copies, such as subset, merge and RenameCells.

pixelatorR.verbose Whether to print messages to the console.

Verbosity

Verbosity can be turned off globally for all pixelatorR function by setting options(pixelatorR.verbose = FALSE).

Edgelists stored on disk

When creating Seurat objects with pixelatorR, the graph data will be stored on disk in a directory that can be set with the global option pixelatorR.arrow_outdir. By default, this directory is set to './edgelists' which will be created in the current working directory.

Certain operations, such as subset, merge and RenameCells, will create copies of the graph data in new sub directories within this folder to make sure that each Seurat object that is created has it's own edgelist file(s) associated with it.

A downside with this strategy is that the size of the directory can grow very large if many Seurat objects are created or if the functions mentioned above are called many times. When Seurat objects are overwritten or removed, some directories may no longer be linked to any global variable in the current R session, but these directories will not be removed automatically. pixelatorR offers a utility function called edgelists_directories_clean() to remove them.

To make the user aware of this behavior, a message will be displayed once the first time ReadMPX_Seurat that creates a new edgelist directory is called.

Moreover, when the size of the directory exceeds the size limit set with the global option pixelatorR.arrowdir_maxsize (which is set to 5GB by default), edgelists_directories_clean() will be run automatically. Users will be given a prompt to confirm the cleanup before it is executed. This behavior can be turned off by setting options(pixelatorR.auto_cleanup = TRUE). You can also run edgelist_directories_du() to get the current size of the directory.

See also

Examples

library(dplyr)

# Set edgelist directory to a temporary folder
options(pixelatorR.arrow_outdir = tempdir())

# Check the current size of the edgelist directory
edgelist_directories_du()
#> 1.39M

# List files in the edgelist directory
getOption("pixelatorR.arrow_outdir") %>% fs::dir_ls()
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/5UDHo-2024-04-02-114046
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/ST5VC-2024-04-02-114046
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/bslib-b2b014205f363e69651141537d9e56af
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/d1Ihl-2024-04-02-114046
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/downlit
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca91d048560
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca92bb67dfc
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca95a0941cf
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca95c97e0a7
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca9771e236e
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca97883bac4
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/file17ca97f08409c
#> /var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpBhxvH3/sorted_components

# Increase maximum allowed size of edgelist directory to 5GB
# This will generate a prompt asking users to cleanup the edgelist directory
# if the total size exceeds 50GB
options(pixelatorR.arrowdir_maxsize = fs::fs_bytes(5 * 1024^3))

# Turn of prompts for automatic cleanup
options(pixelatorR.auto_cleanup = TRUE)

# Turn off verbosity
options(pixelatorR.verbose = FALSE)