Convert colocalization scores to Assay

Convert colocalization scores to Assay

Usage

ColocalizationScoresToAssay(object, ...)

# S3 method for data.frame
ColocalizationScoresToAssay(
  object,
  values_from = c("pearson_z", "pearson", "pearson_mean", "jaccard_mean", "jaccard",
    "jaccard_z"),
  ...
)

# S3 method for CellGraphAssay
ColocalizationScoresToAssay(
  object,
  values_from = c("pearson_z", "pearson", "pearson_mean", "jaccard_mean", "jaccard",
    "jaccard_z"),
  ...
)

# S3 method for Seurat
ColocalizationScoresToAssay(
  object,
  assay = NULL,
  new_assay = NULL,
  values_from = c("pearson_z", "pearson", "pearson_mean", "jaccard_mean", "jaccard",
    "jaccard_z"),
  ...
)

Arguments

object

An object with colocalization scores

...

Not yet implemented

values_from

What column to pick colocalization scores from. One of "pearson", "pearson_mean", "pearson_z", "jaccard_mean", "jaccard" or "jaccard_z"

assay

Name of the CellGraphAssay to pull polarization scores from

new_assay

Name of the Assay to store the polarization scores in

Behavior

Takes an object with colocalization scores in long format and returns an object with colocalization scores in a wide format. The colocalization score table includes various colocalization scores along with p-values for each pair markers and component.

The wide format is an array-like object with dimensions (markers_1 * marker_2) x components, where each cell is filled with a polarization score. Scores that are missing from the colocalization score table are replaced with 0's.

Different outputs are returned depending on the input object type:

tibble/data.frame: returns a matrix with marker pairs in rows and components in columns

CelGraphAssay: returns an Assay with marker pairs in rows and components in columns

Seurat object: returns the Seurat object with a new Assay with marker pairs in rows and components in columns

As many functions provided in Seurat works on Assay objects, it is sometimes convenient to make this conversion. For instance, if we want to compute a UMAP on the colocalization scores with RunUMAP, we need the values to be formatted in an Assay. This also makes it possible to use various visualization functions such as VlnPlot or FeaturePlor to show the distribution of colocalization scores.

See also

Examples

library(pixelatorR)
library(SeuratObject)
#> Loading required package: sp
#> ‘SeuratObject’ was built with package ‘Matrix’ 1.6.3 but the current
#> version is 1.6.5; it is recomended that you reinstall ‘SeuratObject’ as
#> the ABI for ‘Matrix’ may have changed
#> 
#> Attaching package: ‘SeuratObject’
#> The following object is masked from ‘package:base’:
#> 
#>     intersect

# Load example data as a Seurat object
pxl_file <- system.file("extdata/PBMC_10_cells",
  "Sample01_test.pxl",
  package = "pixelatorR"
)
col_scores <- ReadMPX_colocalization(pxl_file)
#> ℹ Loading item(s) from: /private/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpYc215H/temp_libpath1c5e7b138767/pixelatorR/extdata/PBMC_10_cells/Sample01_test.pxl
#> →   Loading colocalization data
#> ✔ Returning a 'tbl_df' object

# ColocalizationScoresToAssay returns a matrix for a tbl_df
col_scores_mat <- ColocalizationScoresToAssay(col_scores)
col_scores_mat[1:4, 1:5]
#>            RCVCMP0000000 RCVCMP0000002 RCVCMP0000003 RCVCMP0000005
#> ACTB-CD11c      6.876119     -1.028363     -7.339336    -2.8884534
#> ACTB-CD14      -6.224955      0.000000      0.000000     0.3047543
#> CD11c-CD14     -6.658966      0.000000      0.000000    -0.6719115
#> ACTB-CD19       5.399166      0.000000     -3.863839     5.4437994
#>            RCVCMP0000006
#> ACTB-CD11c     -1.611557
#> ACTB-CD14       0.000000
#> CD11c-CD14      0.000000
#> ACTB-CD19       1.072440

# Create a Seurat object
seur <- ReadMPX_Seurat(pxl_file, overwrite = TRUE)
#> ✔ Created a 'Seurat' object with 10 cells and 26 targeted surface proteins

# Fetch CellGraphAssay and create new polarization
# scores Assay
cg_assay <- seur[["mpxCells"]]
class(cg_assay)
#> [1] "CellGraphAssay"
#> attr(,"package")
#> [1] "pixelatorR"
col_assay <- ColocalizationScoresToAssay(cg_assay)
class(col_assay)
#> [1] "Assay"
#> attr(,"package")
#> [1] "SeuratObject"

# Convert colocalization scores within a Seurat object
seur <- ColocalizationScoresToAssay(seur)

# After conversion, we now have a new assay called "colocalization"
seur[["colocalization"]]
#> Assay data with 325 features for 10 cells
#> First 10 features:
#>  ACTB-CD11c, ACTB-CD14, CD11c-CD14, ACTB-CD19, CD11c-CD19, CD14-CD19,
#> ACTB-CD20, CD11c-CD20, CD14-CD20, CD19-CD20 

# Switch default assay to polarization
DefaultAssay(seur) <- "colocalization"

# Visualize colocalization scores with Seurat
# VlnPlot(seur, features = "CD19") +
#   ggplot2::labs(y = "Colocalization score")