Load CellGraphs

Loads a list of CellGraph object into the "cellgraphs" slot, provided that an arrow_data object exists in the "arrow_data" slot of a CellGraphAssay. If the arrow_data is missing or the connection to it is dead, the method will attempt to load the edgelist from the arrow_dir stored in the "arrow_dir" slot of the CellGraphAssay.

Usage

LoadCellGraphs(object, ...)

# S3 method for FileSystemDataset
LoadCellGraphs(
  object,
  cells,
  load_as = c("bipartite", "Anode", "linegraph"),
  add_marker_counts = TRUE,
  chunk_length = 10,
  verbose = TRUE,
  ...
)

# S3 method for CellGraphAssay
LoadCellGraphs(
  object,
  cells = colnames(object),
  load_as = c("bipartite", "Anode", "linegraph"),
  add_marker_counts = TRUE,
  force = FALSE,
  chunk_length = 10,
  verbose = TRUE,
  ...
)

# S3 method for Seurat
LoadCellGraphs(
  object,
  assay = NULL,
  cells = colnames(object),
  load_as = c("bipartite", "Anode", "linegraph"),
  add_marker_counts = TRUE,
  force = FALSE,
  chunk_length = 10,
  verbose = TRUE,
  ...
)

Arguments

object

A Seurat object or an CellGraphAssay object

...

Parameters passed to other methods

cells

A character vector of cell names to load CellGraphs for

load_as

Choose how the cell graph should be represented (see details below)

add_marker_counts

Should marker counts be added to the CellGraph objects?

chunk_length

Length of chunks used to load CellGraphs from edge list. Smaller chunks will likely take longer time to load, but decreases memory usage.

verbose

Print messages

force

Force load graph(s) if they are already loaded

assay

Assay name

Value

An object with a list of CellGraph objects

Details

One can specify cells to load by their names using the cells parameter. Only the cells specified will be loaded into the object. Note that the cell graphs are large and will consume a lot of memory. For this reason, it is advisable to load CellGraphs only for cells of interest.

Graphs can be loaded as one of 'bipartite', 'Anode' or 'linegraph'. See details about each of these graph representations in the sections below.

Bi-partite graph

In the bi-partite graph, edges can only go from a upia to a upib. The bi-partite graph is first collapsed from a multigraph to a simple graph by aggregating marker counts for each upia/upib combination. For visualization of marker counts on the graph, it is often convenient to project values on the nodes; however, the marker counts are not available in the nodes in the bi-partite graph. To circumvent this issue, node counts are calculated by aggregating its edge counts. This means that the total marker count will be inflated.

A node projected-graph

In the A node projected-graph, pairs of upias that share a upib are connected with an edge. The (upia) node counts are calculated by aggregating counts across all edges associated with the A nodes. The number of nodes in the A node projected-graph is substantially lower than the bi-partite graph.

Line graph

Starting with a bi-partite graph, a node is placed on each edge. Then, an edge is drawn between each pair of adjacent nodes. The number of nodes in the linegraph is equivalent to the number of edges in the bi-partite graph. However, the number of edges is substantially larger. One major advantage with the linegraph is that the node counts represent the actual raw data, which is not the case for the bi-partite graph and the A node projected-graph. On the down side, linegraphs are much larger and tends to slow down layout computations and visualizations.