Plot 3D graph layouts

Plot a 3D component graph layout computed with ComputeLayout and color nodes by a marker.

Usage

Plot3DGraph(
  object,
  cell_id,
  marker = NULL,
  assay = NULL,
  layout_method = c("pmds", "wpmds", "fr", "kk", "drl"),
  project = FALSE,
  aspectmode = c("data", "cube"),
  colors = c("lightgrey", "mistyrose", "red", "darkred"),
  showgrid = TRUE,
  log_scale = TRUE,
  node_size = 2,
  show_Bnodes = FALSE,
  ...
)

Arguments

object

A Seurat object

cell_id

ID of component to visualize

marker

Name of marker to color the nodes by

assay

Name of assay to pull data from

layout_method

Select appropriate layout previously computed with ComputeLayout

project

Project the nodes onto a sphere. Default FALSE

aspectmode

Set aspect ratio to one of "data" or "cube". If "cube", this scene's axes are drawn as a cube, regardless of the axes' ranges. If "data", this scene's axes are drawn in proportion with the axes' ranges.

Default "data"

colors

Color the nodes expressing a marker. Must be a character vector with at least two color names.

showgrid

Show the grid lines. Default TRUE

log_scale

Convert node counts to log-scale with logp

node_size

Size of nodes

show_Bnodes

Should B nodes be included in the visualization? This option is only applicable to bipartite graphs.

...

Additional parameters

Value

A interactive 3D plot of a component graph layout as a plotly object

Examples

library(pixelatorR)
pxl_file <- system.file("extdata/PBMC_10_cells",
                        "Sample01_test.pxl",
                        package = "pixelatorR")

seur <- ReadMPX_Seurat(pxl_file, overwrite = TRUE)
#> ✔ Created a 'Seurat' object with 10 cells and 26 targeted surface proteins
seur <- LoadCellGraphs(seur, cells = colnames(seur)[1:10])
#> →   Loading 10 edgelist(s) as bipartite graph(s)
#> ✔ Successfully loaded 10 CellGraph object(s).
seur[["mpxCells"]] <- KeepLargestComponent(seur[["mpxCells"]])
#> ℹ Keeping largest component for 10 graphs
#> ℹ Removed 24 out of 3549 nodes
#> ℹ Removed 227 out of 1649 nodes
#> ℹ Removed 83 out of 2252 nodes
#> ℹ Removed 231 out of 3644 nodes
#> ℹ Removed 348 out of 4401 nodes
#> ℹ Removed 88 out of 1305 nodes
#> ℹ Removed 70 out of 1547 nodes
#> ℹ Removed 374 out of 2158 nodes
#> ℹ Removed 1131 out of 1330 nodes
#> ℹ Removed 101 out of 5129 nodes
seur <- ComputeLayout(seur, layout_method = "pmds", dim = 3)
#> ℹ Computing layouts for 10 graphs

Plot3DGraph(seur, cell_id = colnames(seur)[1], marker = "HLA-ABC")