Read a pixel data item

Function that unzips a .pxl file into a temporary folder and reads any number of items (not anndata) as specified in items. If more than one item is fetched, the output is a list of data frames. Otherwise the output is a single tbl_df.

Usage

ReadMPX_item(
  filename,
  items = c("colocalization", "polarization", "edgelist"),
  verbose = TRUE
)

ReadMPX_polarization(filename, verbose = TRUE)

ReadMPX_colocalization(filename, verbose = TRUE)

ReadMPX_edgelist(filename, verbose = TRUE)

Arguments

filename

Path to a .pxl file

items

One of "colocalization", "polarization", "edgelist"

verbose

Print messages

Value

Either an object of class tbl_df or a list of tbl_df

objects if multiple items are selected

See also

Examples

library(pixelatorR)

# Load example data
pxl_file <- system.file("extdata/PBMC_10_cells",
                        "Sample01_test.pxl",
                        package = "pixelatorR")
polarization <- ReadMPX_item(pxl_file, items = "polarization")
#> ℹ Loading item(s) from: /private/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpHmlprD/temp_libpath148616148d180/pixelatorR/extdata/PBMC_10_cells/Sample01_test.pxl
#> →   Loading polarization data
#> ✔ Returning a 'tbl_df' object
polarization
#> # A tibble: 225 × 6
#>     morans_i morans_p_value morans_p_adjusted morans_z marker component    
#>        <dbl>          <dbl>             <dbl>    <dbl> <chr>  <chr>        
#>  1 -0.000468        0.977              1.00    -0.0288 ACTB   RCVCMP0000000
#>  2  0.00233         0.521              1.00     0.642  CD11c  RCVCMP0000000
#>  3  0.0118          0.00143            0.0183   3.19   CD14   RCVCMP0000000
#>  4 -0.00106         0.875              1.00    -0.157  CD19   RCVCMP0000000
#>  5  0.00455         0.289              0.974    1.06   CD20   RCVCMP0000000
#>  6 -0.00274         0.605              1.00    -0.517  CD22   RCVCMP0000000
#>  7  0.00708         0.108              0.574    1.61   CD26   RCVCMP0000000
#>  8 -0.000123        0.962              1.00     0.0474 CD29   RCVCMP0000000
#>  9  0.000420        0.869              1.00     0.164  CD32   RCVCMP0000000
#> 10 -0.00272         0.568              1.00    -0.571  CD328  RCVCMP0000000
#> # ℹ 215 more rows

# Alternative 2
polarization <- ReadMPX_polarization(pxl_file)
#> ℹ Loading item(s) from: /private/var/folders/gw/bdcqhnvs0m9gs_mq8n51jtbc0000gn/T/RtmpHmlprD/temp_libpath148616148d180/pixelatorR/extdata/PBMC_10_cells/Sample01_test.pxl
#> →   Loading polarization data
#> ✔ Returning a 'tbl_df' object